reorganize nycmesh package
